Migrate CreateApiRuleHeader to a separate component
