+++ core-ui/src/components/ApiRules/CreateApiRule/CreateApiRuleHeader/CreateApiRuleHeader.scss
@@ -1,0 +1,11 @@
+.copiable-url {
+  div {
+    position: relative;
+    button {
+      position: absolute;
+      transform: translateY(-50%);
+      top: 50%;
+      margin-left: 4px;
+    }
+  }
+}
